Added telegraph audio
Also for a introduction, attack speeds of 1.5, 3, 5 seem to be fine. There is no need to worry about multiple attacks at once
The speed of the first one could be changed to be lower(1 or less) as the boss gets less health so the player needs to multitask a bit.
This would make the arrow/firefly distractors more useful as they can guide you to the closest one.