Added VR support
Things to note:
There is a bug in Unity that currently results in heavy performance hits with particle systems in VR. This is rather problematic for the fireflies.

Also: Asynchronous/interleaved reprojection can decrease the framerate of the game when turned on. (To 45 assuming the hardware cant keep up with 90)
Turning them off in SteamVR makes things a lot more smooth, but it might(?) make you feel more sick. This was the case for me at least, but it might also just be a result of sickness buildup from testing.